+++ segmentation/classical.py
@@ -20,7 +20,7 @@ def otsu_thresholding(image, k, mode="color"):
         return segmented_image,thresholds
     else:
         thresholds = []
-        segmented_image = np.zeros_like(image, dtype=np.float32)
+        segmented_image = np.zeros_like(image, dtype=np.uint8)
         for channel in range(3):
             channel_thresholds = threshold_multiotsu(image[:, :, channel], classes=k)
             segmented_image[:, :, channel] = apply_thresholds(image[:, :, channel], channel_thresholds)
@@ -37,10 +37,10 @@ def kmeans_thresholding(image, k, mode="color"):
         segmented_image = cluster_centers[y_pred].reshape(image.shape).astype(np.uint8)
         return segmented_image, sorted(cluster_centers.tolist())
     else:
-        segmented_image = np.zeros_like(image, dtype=np.float32)
+        segmented_image = np.zeros_like(image, dtype=np.uint8)
         thresholds = []
         for channel in range(3):
-            x = image[:, :, channel].flatten().reshape(-1, 1).astype(np.float32)
+            x = image[:, :, channel].flatten().reshape(-1, 1).astype(np.uint8)
             kmeans = KMeans(n_clusters=k, random_state=42,n_init=10)
             y_pred = kmeans.fit_predict(x)
             cluster_centers = kmeans.cluster_centers_.flatten()
